+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 3.0.6
+- Fixed issue when deleteting combatant in Combat Traker before combatant begins.
+- At Post-Segment-12 all active combatants Take a Recovery.
+- Stun status is cleared at the beginning of phase.
+- Initial Combat Skill Levels (CSL) support.  OCV is added to attacks.  Simple +1DC. DCV (like all DCV modifiers) is shown but not currently implemented. [#166](https://github.com/dmdorman/hero6e-foundryvtt/issues/166)
+
 # Version 3.0.5
 - Initial DRAIN support.
 - Changing PC/NPC actor type moved to sheet header.  Also can be changed in the context menu of the actor sidebar. Fixes [#170](https://github.com/dmdorman/hero6e-foundryvtt/issues/170).